requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.59.6
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
Telethon==1.38.1
tqdm==4.67.1
twilio==9.4.2
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
